UP DTO DAL

+++ DTO/KetQuaDTO.cs
@@ -11,14 +11,14 @@ namespace DTO
         public int MaKetQua {  get; set; }
         public int MaDe {  get; set; }
         public int MaNguoiDung { get; set; } //MaSV
-        public int Diem {  get; set; }
+        public decimal Diem {  get; set; }
         public int SoCauDung { get; set; }
         public int SoCauSai {  get; set; }
         public int TrangThai { get; set; }
         public int is_delete { get; set; }
         public KetQuaDTO() { }
 
-        public KetQuaDTO(int maKetQua, int maDe, int maNguoiDung, int diem, int soCauDung, int soCauSai, int trangThai, int is_delete)
+        public KetQuaDTO(int maKetQua, int maDe, int maNguoiDung, decimal diem, int soCauDung, int soCauSai, int trangThai, int is_delete)
         {
             MaKetQua = maKetQua;
             MaDe = maDe;
